pyproject.toml
[project]
name = "codegraphcontext"
version = "0.1.0"
description = "An MCP server that indexes local code into a graph database to provide context to AI assistants."
authors = [{ name = "Shashank Shekhar Singh", email = "shashankshekharsingh1205@gmail.com" }]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]
dependencies = [
    "neo4j>=5.15.0",
    "watchdog>=3.0.0",
    "requests>=2.31.0",
    "stdlibs>=2023.11.18",
    "typer[all]>=0.9.0",
    "rich>=13.7.0",
    "inquirerpy>=0.3.4",
    "python-dotenv>=1.0.0",
]

[project.urls]
"Homepage" = "https://github.com/Shashankss1205/CodeGraphContext"
"Bug Tracker" = "https://github.com/Shashankss1205/CodeGraphContext/issues"

[project.scripts]
cgc = "codegraphcontext.cli.main:app"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.11.0",
]

[tool.setuptools]
packages = ["codegraphcontext"]
package-dir = { "" = "src" }